% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{c.patternogram}
\alias{c.patternogram}
\title{Combine Patternogram Objects}
\usage{
\method{c}{patternogram}(..., ids = NULL)
}
\arguments{
\item{...}{One or more objects of class \code{patternogram}.}

\item{ids}{Optional character vector of identifiers to assign to each input object. If \code{NULL}, identifiers are generated automatically.}
}
\value{
A \code{patternogram} object containing all input objects combined row-wise, with an added \code{id} column indicating their origin
}
\description{
This function provides a method for \code{c()} that combines multiple
\code{patternogram} objects into a single object, binding them by rows.
}
\examples{
r = terra::rast(system.file("ex/elev.tif", package = "terra"))
r1 = r[r < 340, drop = FALSE]
r2 = r[r >= 340, drop = FALSE]
pr1 = patternogram(r1)
pr2 = patternogram(r2)
pr_all = c(pr1, pr2, ids = c("low", "high"))
plot(pr_all)
}
