% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternogram-package.R
\docType{package}
\name{patternogram-package}
\alias{patternogram-package}
\title{patternogram: A Visualization of Spatial Autocorrelation}
\description{
The patternogram represents the average dissimilarity between the set of values in the given locations as a function of spatial distances between them. It is allows to: detect spatial autocorrelation for many data structures; compare spatial autocorrelation of variables over time; explore spatial autocorrelations of predictors in spatial machine learning models; analyze spatial autocorrelation of categorical spatial patterns.
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
