% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{plot.patternogram}
\alias{plot.patternogram}
\title{Plot a patternogram}
\usage{
\method{plot}{patternogram}(x, ...)
}
\arguments{
\item{x}{An output of the \code{patternogram()} function}

\item{...}{Additional arguments to \code{autoplot()}}
}
\description{
Plot a patternogram
}
