% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_patternogram.R
\name{rescale_patternogram}
\alias{rescale_patternogram}
\title{Rescale Values in a Patternogram}
\usage{
rescale_patternogram(x, method = c("minmax"), by = "id", ...)
}
\arguments{
\item{x}{A patternogram object}

\item{method}{Scaling method, currently only "minmax" is implemented}

\item{by}{Optional grouping variable (e.g., "id" or "group") to apply scaling within groups}

\item{...}{Additional arguments for future extensions}
}
\value{
A patternogram object with rescaled dissimilarity values (0 to 1 for "minmax" method)
}
\description{
It rescales the dissimilarity values in a patternogram object to a specified range.
This allows for easier comparison and interpretation of various patternograms.
}
\examples{
r = terra::rast(system.file("ex/elev.tif", package = "terra"))
pr = patternogram(r)
pr
pr_scaled = rescale_patternogram(pr)
pr_scaled
plot(pr_scaled)
}
