% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.patternogram}
\alias{autoplot.patternogram}
\title{Plot a patternogram}
\usage{
\method{autoplot}{patternogram}(plot_data, ...)
}
\arguments{
\item{plot_data}{An output of the \code{patternogram()} function}

\item{...}{Additional arguments to \code{autoplot()}}
}
\description{
Plot a patternogram
}
