% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternogram.R
\name{patternogram}
\alias{patternogram}
\title{Create a Patternogram or a Patternogram Cloud}
\usage{
patternogram(
  x,
  cutoff,
  width = cutoff/15,
  dist_fun = "euclidean",
  sample_size = 500,
  cloud = FALSE,
  target = NULL,
  ...
)
}
\arguments{
\item{x}{A raster object of class SpatRaster (terra) or a point vector object of class sf (sf)}

\item{cutoff}{Spatial distance up to which point pairs are included in patternogram estimates;
by default: a square root of the raster area}

\item{width}{The width of subsequent distance intervals for which data point pairs are grouped for patternogram estimates}

\item{dist_fun}{Distance measure used. This function uses the \code{philentropy::distance()} function (run \code{philentropy::getDistMethods()} to find possible distance measures)}

\item{sample_size}{Only used when \code{x} is raster. Proportion of the cells inside of each region to be used in calculations. Value between 0 and 1. It is also possible to specify an integer larger than 1, in which case the specified number of cells of each region will be used in calculations.}

\item{cloud}{Logical; if TRUE, calculate the patternogram cloud}

\item{target}{Additional argument allowing to calculate separate estimates for different categories or ranges of values}

\item{...}{Additional arguments for \code{base::cut()}}
}
\value{
A tibble of the patternogram class with columns, such as (a) np - the number of point pairs in this estimate, (b) dist - the middle of the distance interval used for each estimate, (c) dissimilarity - the dissimilarity estimate, (d) (optional) target
}
\description{
It creates a patternogram or a patternogram cloud. The function takes a raster object of class \code{SpatRaster} or a point vector object of class \code{sf} as input, and calculates the dissimilarity between values of pairs of points given the distances between them. The output of this function is a tibble of the patternogram class that can be visualized with the \code{plot()} and \code{autoplot()} functions.
}
\examples{
r = terra::rast(system.file("ex/elev.tif", package = "terra"))
pr = patternogram(r)
pr
plot(pr)
}
