% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternogram.R
\name{patternogram}
\alias{patternogram}
\title{Title}
\usage{
patternogram(
  x,
  cutoff,
  width = cutoff/15,
  dist_fun = "euclidean",
  sample_size = 100,
  approach,
  cloud = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
